Slim down launcher
