+++ dev/dev.cfg
@@ -5,10 +5,6 @@ FEED_ID = "tag:example.com,2005:komorebi"
 # Avoid generating session cookies.
 SECRET_KEY = None
 
-# Defaults that would need to be changed for production.
-APPLICATION_ROOT = "/"
-PREFERRED_URL_SCHEME = "http"
-
 # Application configuration.
 DB_PATH = "db.sqlite"
 HTPASSWD_PATH = "dev/dev.htpasswd"
